fix: unify diarization module by standardizing naming to speaker_diarization
